fix(ci): retire review gate and simplify PR checks
