+++ quality/quarantine/review-gate/workflows/merge-group-recheck.yml
@@ -1,3 +1,4 @@
+# Quarantined after the 2026-07-30 SEV-1. This file is not an active workflow.
 # Re-verify queue-readiness INSIDE the merge group.
 #
 # Why this exists: arming covers the path into the queue, and the automerge
